Add cpp07.ex01 to repo

+++ Rank05/CPP/CPP_07/ex01/include/iter.hpp
@@ -2,7 +2,7 @@
 #define ITER_HPP
 
 template <typename T>
-void	iter(T *array[], int const length, void (*f)(T))
+void	iter(T *array[], size_t const length, void (*f)(T))
 {
 	for (int i = 0; i < length; i++)
 	{
